app/requirements.txt
python-dotenv
pandas
regex
vk_api
tqdm==4.66.1
sentencepiece==0.1.99
transformers==4.36.0
bitsandbytes==0.41.3.post2
peft==0.7.1
accelerate==0.25.0
auto-gptq==0.5.1
nvgpu==0.10.0
pynvml==11.5.0
protobuf==4.25.1
